ReviewStatistics: Change to 1 d.p. display
Edits:
- Change JsonAdaptedDeck:toModelType() to read doubles instead.
- Remove redundant methods in JsonAdaptedDeck class.
- Change IndividualFlashcard::displayStatistics() to generate statistics in double instead of integer.
- Update DeckCard class to check for if deck statistic string is the default instead of comparing to a value.
- Change method signature for updateDeckPerformanceScore() method.
- Change default value of Deck class resultStatistics variable.

+++ src/test/java/seedu/flashnotes/logic/commands/AddDeckCommandTest.java
@@ -258,7 +258,7 @@ public class AddDeckCommandTest {
         }
 
         @Override
-        public void updateDeckPerformanceScore(Integer reviewScore, String deckName) {
+        public void updateDeckPerformanceScore(Double reviewScore, String deckName) {
             throw new AssertionError("This method should not be called.");
         }
