+++ src/test/java/seedu/flashnotes/logic/commands/DeleteDeckCommandTest.java
@@ -243,7 +243,7 @@ public class DeleteDeckCommandTest {
         }
 
         @Override
-        public void updateDeckPerformanceScore(Integer reviewScore, String deckName) {
+        public void updateDeckPerformanceScore(Double reviewScore, String deckName) {
             throw new AssertionError("This method should not be called.");
         }
